Large refactoring in order to start splitting the syntactiacl operations that are plutus-independent (#51)
* Massive refactorings in an attempt to split what is PlutusIR
independent from what is not. This went through defining a
new class, called LanguageDef; which we have to pass around
everywhere. Using a 'PrtTerm lang' signifies that that function
is fully independent from PlutusIR.

I finished doing the refactoring for the library part of pirouette;
will do the executable tomorrow.

* Pirouette builds!

* get rid of warning

* Update src/Pirouette/Term/Syntax/Base.hs