Large refactoring on the SymbolicEval and SMT components (#52)
* Massive refactorings in an attempt to split what is PlutusIR
independent from what is not. This went through defining a
new class, called LanguageDef; which we have to pass around
everywhere. Using a 'PrtTerm lang' signifies that that function
is fully independent from PlutusIR.

I finished doing the refactoring for the library part of pirouette;
will do the executable tomorrow.

* Pirouette builds!

* get rid of warning

* Start a large refactoring effort, splitting up the symbolic evaluation
into smaller chunks. A key component introduced here is the usage of
'metavariables' to represent symbolic variables, instead of relying on
bound vars and having to deal with renamings and stuff

* The heard of symbolic evaluation is starting to take a cleaner shape.
I don't like ListT very much and might very well end up playing with
returning an `Alternative` explicitely

* Alright; There's some structure coming out of this

* Comment about changing the 'Path' type

* It runs. Not correctly though...

* Ok... we're learning the right things, but maybe lets not propagate on bind

* I'm losing my constraints somewhere

* That's much simpler. I was overcomplicating my life

* Large refactoring moving Plutus stuff out of Term namespace and moving symbolic eval into Term namespace

* Removed old code that was already translated; added a note on where to
plug unification in and inlined an auxiliar function for better readability

* Started refactoring PlutusIR out of the SMT folder

* SMT module is now compiling; can still use a refactoring

* Turn SMT back on!!

* Maybe bound-vars should be forbidden, not meta

* Rely on a nicer unification-ish that creates less symbolic variables

* Incomplete pattern matches fixed

* Update comment on zipWithMPlus

* Refactoring the SMT folder

* Massive SMT refactor; made a SolverT monad and started using push and pop instead of redeclaring the whole context

* Only sent a branch to SMT if it hasn't been validated before

* Make smt translation call `fail` instead of `error`, which in turn
enable us to catch potential problems.

* Incorporate GR's comments