Fix mono matches (1) (#128)
* Change the type of the functional tyargs to matchers to closure types

Those are to be defunctionalized in the match branches, so update the
type accordingly.

* `fixPrefixType` is not the best name

* Defun just the datatype-related tyargs to matchers

* Defunctionalization of dtors is not the responsibility of defunTypes

* Added more tests showing more bugs coming from monomorphization

* Fix monomorphization

* Forgot to turn on the rest of the tests

* We need to defunctionalize nested things too; made the enging pick
these up, but need to tweag the defunctionalization of destructors

* Getting close, now we need to just call the right apply on our example

* I was too eager, have to leave some TyFun's for vanilla
defunctionalization to kick in. Now the _Apply function is not being generated

* I think my hacks from before are doomed to fail. By the time we get to
generate the apply functions, we never generated a HofArgInfo for the
nested calls.

This function has to have two steps: defun nested types,generating
their proper [HofArgInfo]; then defun types. With a little luck, that
should also handle constructors alright!

* I think I got the errors out

* Found another case that breaks the defunctionalizer... will mark it as expectedFailure for now, we should rewrite that module at some point

Co-authored-by: Victor Miraldo <victor.miraldo@tweag.io>